Add tests for remaining hooks
